Move bike routing test to other routing tests
