+++ src/test/scala/beam/router/TimeDependentRoutingSpec.scala
@@ -128,6 +128,27 @@ class TimeDependentRoutingSpec
       // presumably because the first/last edge are not travelled (in R5, trip starts on a "split")
     }
 
+    "also for bikes" in {
+      val leg = BeamLeg(
+        3000,
+        BeamMode.BIKE,
+        0,
+        BeamPath(
+          Vector(143, 60, 58, 62, 80, 74, 68, 154),
+          Vector(),
+          None,
+          SpaceTime(166321.9, 1568.87, 3000),
+          SpaceTime(167138.4, 1117, 3000),
+          0.0
+        )
+      )
+      router ! EmbodyWithCurrentTravelTime(leg, Id.createVehicleId(1), Id.create("Bicycle", classOf[BeamVehicleType]))
+      val response = expectMsgType[RoutingResponse]
+      assert(response.itineraries.head.beamLegs().head.duration == 285)
+      // R5 travel time, but less than what's in R5's routing response (see vv),
+      // presumably because the first/last edge are not travelled (in R5, trip starts on a "split")
+    }
+
     "take given link traversal times into account" in {
       router ! RoutingRequest(
         origin,
